consolidate streaming
