+++ .vscode/launch.json
@@ -12,8 +12,8 @@
       "cwd": "${workspaceFolder}/extensions/vscode",
       "args": [
         // Pass a directory to manually test in
-        "/Users/dallin/Documents/code/continuedev/continue-extras/dlt-example",
-        "/Users/dallin/Documents/code/continuedev/continue-extras/dlt-example/start.py",
+        "${workspaceFolder}/manual-testing-sandbox",
+        "${workspaceFolder}/manual-testing-sandbox/test.js",
         "--extensionDevelopmentPath=${workspaceFolder}/extensions/vscode"
       ],
       "pauseForSourceMap": false,
