+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/constants/MessageTypes.kt
@@ -103,7 +103,6 @@ class MessageTypes {
             "autocomplete/complete",
             "autocomplete/cancel",
             "autocomplete/accept",
-            "command/run",
             "tts/kill",
             "llm/complete",
             "llm/streamComplete",
