+++ gui/src/redux/thunks/index.ts
@@ -1,8 +1,7 @@
 export { exitEditMode } from "./exitEditMode";
 export { gatherContext } from "./gatherContext";
-export { streamThunkWrapper } from "./streamThunkWrapper";
 export { resetStateForNewMessage } from "./resetStateForNewMessage";
 export { streamNormalInput } from "./streamNormalInput";
 export { streamResponseThunk } from "./streamResponse";
 export { streamResponseAfterToolCall } from "./streamResponseAfterToolCall";
-export { streamSlashCommand } from "./streamSlashCommand";
+export { streamThunkWrapper } from "./streamThunkWrapper";
