+++ gui/src/redux/thunks/streamNormalInput.ts
@@ -1,5 +1,7 @@
 import { createAsyncThunk, unwrapResult } from "@reduxjs/toolkit";
-import { ChatMessage, PromptLog } from "core";
+import { ChatMessage } from "core";
+import { modelSupportsTools } from "core/llm/autodetect";
+import { ToCoreProtocol } from "core/protocol";
 import { selectCurrentToolCall } from "../selectors/selectCurrentToolCall";
 import { selectDefaultModel } from "../slices/configSlice";
 import {
@@ -10,98 +12,109 @@ import {
 } from "../slices/sessionSlice";
 import { ThunkApiType } from "../store";
 import { callTool } from "./callTool";
-import { modelSupportsTools } from "core/llm/autodetect";
 
 export const streamNormalInput = createAsyncThunk<
   void,
-  ChatMessage[],
+  {
+    messages: ChatMessage[];
+    legacySlashCommandData?: ToCoreProtocol["llm/streamChat"][0]["legacySlashCommandData"];
+  },
   ThunkApiType
->("chat/streamNormalInput", async (messages, { dispatch, extra, getState }) => {
-  // Gather state
-  const state = getState();
-  const defaultModel = selectDefaultModel(state);
-  const toolSettings = state.ui.toolSettings;
-  const streamAborter = state.session.streamAborter;
-  const useTools = state.ui.useTools;
-  if (!defaultModel) {
-    throw new Error("Default model not defined");
-  }
-
-  const includeTools =
-    useTools &&
-    modelSupportsTools(defaultModel) &&
-    state.session.mode === "chat";
-
-  // Send request
-  const gen = extra.ideMessenger.llmStreamChat(
-    defaultModel.title,
-    streamAborter.signal,
-    messages,
-    includeTools
-      ? {
-          tools: state.config.config.tools.filter(
-            (tool) => toolSettings[tool.function.name] !== "disabled",
-          ),
-        }
-      : {},
-  );
-
-  // Stream response
-  let next = await gen.next();
-  while (!next.done) {
-    if (!getState().session.isStreaming) {
-      dispatch(abortStream());
-      break;
+>(
+  "chat/streamNormalInput",
+  async (
+    { messages, legacySlashCommandData },
+    { dispatch, extra, getState },
+  ) => {
+    // Gather state
+    const state = getState();
+    const defaultModel = selectDefaultModel(state);
+    const toolSettings = state.ui.toolSettings;
+    const streamAborter = state.session.streamAborter;
+    const useTools = state.ui.useTools;
+    if (!defaultModel) {
+      throw new Error("Default model not defined");
     }
 
-    dispatch(streamUpdate(next.value));
-    next = await gen.next();
-  }
+    const includeTools =
+      useTools &&
+      modelSupportsTools(defaultModel) &&
+      state.session.mode === "chat";
 
-  // Attach prompt log
-  if (next.done && next.value) {
-    dispatch(addPromptCompletionPair([next.value]));
+    // Send request
+    const gen = extra.ideMessenger.llmStreamChat(
+      {
+        completionOptions: includeTools
+          ? {
+              tools: state.config.config.tools.filter(
+                (tool) => toolSettings[tool.function.name] !== "disabled",
+              ),
+            }
+          : {},
+        title: defaultModel.title,
+        messages,
+        legacySlashCommandData,
+      },
+      streamAborter.signal,
+    );
 
-    try {
-      if (state.session.mode === "chat") {
-        extra.ideMessenger.post("devdata/log", {
-          name: "chatInteraction",
-          data: {
-            prompt: next.value.prompt,
-            completion: next.value.completion,
-            modelProvider: defaultModel.provider,
-            modelTitle: defaultModel.title,
-            sessionId: state.session.id,
-          },
-        });
+    // Stream response
+    let next = await gen.next();
+    while (!next.done) {
+      if (!getState().session.isStreaming) {
+        dispatch(abortStream());
+        break;
       }
-      // else if (state.session.mode === "edit") {
-      //   extra.ideMessenger.post("devdata/log", {
-      //     name: "editInteraction",
-      //     data: {
-      //       prompt: next.value.prompt,
-      //       completion: next.value.completion,
-      //       modelProvider: defaultModel.provider,
-      //       modelTitle: defaultModel.title,
-      //     },
-      //   });
-      // }
-    } catch (e) {
-      console.error("Failed to send dev data interaction log", e);
-    }
-  }
 
-  // If it's a tool call that is automatically accepted, we should call it
-  const toolCallState = selectCurrentToolCall(getState());
-  if (toolCallState) {
-    dispatch(setToolGenerated());
-
-    if (
-      toolSettings[toolCallState.toolCall.function.name] ===
-      "allowedWithoutPermission"
-    ) {
-      const response = await dispatch(callTool());
-      unwrapResult(response);
+      dispatch(streamUpdate(next.value));
+      next = await gen.next();
     }
-  }
-});
+
+    // Attach prompt log
+    if (next.done && next.value) {
+      dispatch(addPromptCompletionPair([next.value]));
+
+      try {
+        if (state.session.mode === "chat") {
+          extra.ideMessenger.post("devdata/log", {
+            name: "chatInteraction",
+            data: {
+              prompt: next.value.prompt,
+              completion: next.value.completion,
+              modelProvider: defaultModel.provider,
+              modelTitle: defaultModel.title,
+              sessionId: state.session.id,
+            },
+          });
+        }
+        // else if (state.session.mode === "edit") {
+        //   extra.ideMessenger.post("devdata/log", {
+        //     name: "editInteraction",
+        //     data: {
+        //       prompt: next.value.prompt,
+        //       completion: next.value.completion,
+        //       modelProvider: defaultModel.provider,
+        //       modelTitle: defaultModel.title,
+        //     },
+        //   });
+        // }
+      } catch (e) {
+        console.error("Failed to send dev data interaction log", e);
+      }
+    }
+
+    // If it's a tool call that is automatically accepted, we should call it
+    const toolCallState = selectCurrentToolCall(getState());
+    if (toolCallState) {
+      dispatch(setToolGenerated());
+
+      if (
+        toolSettings[toolCallState.toolCall.function.name] ===
+        "allowedWithoutPermission"
+      ) {
+        const response = await dispatch(callTool());
+        unwrapResult(response);
+      }
+    }
+  },
+);
